isolate PatcherViewport
